Adopted "parent" names from ThePornDB
ThePornDB uses two names for each performer: a "site" name that applies only to that site, and a "parent" name that applies across sites.  This edit revises the script to add performers only if a "parent" name exists.  If not, we assume the data is not reliable to justify adding to stash.