Create Dockerfile
Creates a Docker Image to put the script in to and automatically run it so there is no question on the dependencies. Created this after running into a conflict where Pillow v7 was found to be incompatible with Python 3.9 and the script failed to run with Pillow v8 which is compatible with Python 3.9.